backend/qa_service/requirements.txt
Flask
faiss-cpu
sentence-transformers
PyPDF2